pyproject.toml
[project]
name = "trenerai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "exercise>=0.1.1",
    "fastapi>=0.125.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.5",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
